allow to determine the flavour of the used model to setup the execute-mojo for that model in a way so that proper maven can execute it as well.
still needs each polyglot extension to pass in the options-map to do it right.
fallback is the old behaviour.

biggest issue here is the VERSION of the plugin which is hardcoded - maybe there is a better way to do it or it is just OK since the plugin might see a different release cycle.