Review setup and tests, disabled some specific tests under Hava 11
The 7 disabled tests fail in their test setup, which somehow stopped working with Java 11.
The actual tested code - the plugin - still works regardless of the Java version used to compile (8 or 11) and run (8 or 11).
I tested this manually for now, but adding some maven-invoker-plugin based tests is probably a good idea.