Change context classloader when starting clojure runtime
If running under maven 3.8.4, the clojure extension dies when it's
trying to initialize clojure runtime. Unlike in #117, this does not
happen during compilation of clojure sources, but rather when reading
`pom.clj`.

The fix changes the context classloader of the current thread from
whatever maven now sets it to to a classloader known to have clojure
available. It's copied essentially verbatim from [this SO
answer](https://stackoverflow.com/a/13711141/6564029).