Minimize 'locatePom()'-calls in PolyglotModelManager
For some Mapping implementations calling 'locatePom()' can be a
relatively expensive operation, for example because it includes many
file-system interactions.

Because only the result of the mapping with the highest priority is
returned eventually, 'findPom()' and 'determineFlavourFromPom()' can be
speed-up if Mappings are processed in descending priority-order and
simply return the first encountered result.
Just like the get*() methods do.