Update JRuby and all JRuby/Maven plugins
This is part of work to update the entire JRuby/Maven stack, in
order to update the version of JRuby used and to address API
deprecations at rubygems.org.

This PR updates JRuby to latest (9.4.3.0) and moves all plugins
to their new org.jruby.maven group ID with latest versions.

The maven-tools gem is updated to 1.2.

Part of the fix for torquebox/maven-tools#37.

We will need to coordinate getting this released along with the
maven plugins in the very near term. Because of the rubygems.org
API being shut down next week (the 8th), we are under the gun.

See jruby/mavengem#9, jruby/jruby-maven-plugins#126,
torquebox/maven-tools#38 and jruby/jruby#7872.