Replace signing
We are forced to use takari-lifecycle 2.0.8 to support Java 8 building
this project, but alas, signing using that version is busted.

Hence, we need to skip it, and provide alternate way using
vanilla plugin, as Central requires signatures.

IMHO, this is okay compromise, as signing still happens in
takari-release profile only.

Fixes #274.